+++ src/base/utils/string.h
@@ -41,15 +41,6 @@ class QStringRef;
 
 namespace Utils::String
 {
-    QString fromDouble(double n, int precision);
-
-    int naturalCompare(const QString &left, const QString &right, const Qt::CaseSensitivity caseSensitivity);
-    template <Qt::CaseSensitivity caseSensitivity>
-    bool naturalLessThan(const QString &left, const QString &right)
-    {
-        return (naturalCompare(left, right, caseSensitivity) < 0);
-    }
-
     QString wildcardToRegexPattern(const QString &pattern);
 
     template <typename T>
@@ -71,6 +62,8 @@ namespace Utils::String
     std::optional<double> parseDouble(const QString &string);
 
     QString join(const QVector<QStringRef> &strings, const QString &separator);
+
+    QString fromDouble(double n, int precision);
 
     template <typename T, typename std::enable_if_t<std::is_enum_v<T>, int> = 0>
     QString fromEnum(const T &value)
